chore: moved read functions to deserialize
